requirements.txt
Click==7.0
coverage==5.0.3
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
python-dotenv==0.10.5
SQLAlchemy==1.3.13
Werkzeug==0.16.1
